requirements.txt
Flask==2.1.2
Pillow==9.1.1
pygal==3.0.0
